+++ e2e/kit/src/routes/stores/ssr-loadall-store-without-variables/+page.ts
@@ -1,0 +1,6 @@
+import { load_Hello, loadAll } from '$houdini';
+import type { LoadEvent } from '@sveltejs/kit';
+
+export async function load(event: LoadEvent) {
+  return await loadAll(load_Hello({ event }));
+}