+++ packages/houdini-react/src/plugin/vite.tsx
@@ -54,7 +54,6 @@ export default {
 						assetFileNames: 'assets/[name].js',
 						entryFileNames: '[name].js',
 					},
-					external: ['react-streaming/server'],
 				},
 			}