+++ packages/houdini-react/src/runtime/server/utils/isClientSide.ts
@@ -1,0 +1,5 @@
+export { isClientSide }
+
+function isClientSide() {
+	return typeof window !== 'undefined' && typeof window?.getComputedStyle === 'function'
+}