+++ packages/houdini-react/CHANGELOG.md
@@ -1,8 +1,26 @@
 # houdini-react
 
+## 1.2.27
+
+### Dependency Changes
+
+-   Updated dependencies []:
+    -   houdini@1.2.27
+
+## 1.2.26
+
+### 🐛 Fixes
+
+-   [#1205](https://github.com/HoudiniGraphql/houdini/pull/1205) [`e05f1c25`](https://github.com/HoudiniGraphql/houdini/commit/e05f1c250edb48722ee45c2a62de72f4a19d3357) Thanks [@AlecAivazis](https://github.com/AlecAivazis)! - Fix invalid import during dev
+
+### Dependency Changes
+
+-   Updated dependencies []:
+    -   houdini@1.2.26
+
 ## 1.2.25
 
-### Patch Changes
+### Dependency Changes
 
 -   [`0c81de6c`](https://github.com/HoudiniGraphql/houdini/commit/0c81de6ca4a266a646948a09bdc83428379504f3) Thanks [@AlecAivazis](https://github.com/AlecAivazis)! - Improve integration with vite dev server
 
@@ -11,7 +29,7 @@
 
 ## 1.2.24
 
-### Patch Changes
+### Dependency Changes
 
 -   [`d5559503`](https://github.com/HoudiniGraphql/houdini/commit/d55595038b95ccef3e19860e10af63a014a039fa) Thanks [@AlecAivazis](https://github.com/AlecAivazis)! - Fix base url in dev mod